create card

+++ src/main/java/com/ais/project/repo/CardRepository.java
@@ -3,5 +3,5 @@ package com.ais.project.repo;
 import com.ais.project.models.Card;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CardRepository extends CrudRepository<Card, Integer> {
+public interface CardRepository extends CrudRepository<Card, Long> {
 }
